requirements.txt
openai
termcolor
python-dotenv
requests
scikit-learn